Split test_sqlitecachedbhandler.py into multiple files
